backend/main.py
"""
MisterHR Backend API - Multi-Agent LangChain Platform
Core FastAPI application with AI agent orchestration for resume optimization and candidate screening.
"""

from fastapi import FastAPI, HTTPException, status
from fastapi.middleware.cors import CORSMiddleware
import uvicorn
from pydantic import BaseModel
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Initialize FastAPI app
app = FastAPI(
    title="MisterHR API",
    description="AI-powered platform for resume optimization and candidate screening using multi-agent LangChain framework",
    version="1.0.0",
    docs_url="/docs"
)

# Configure CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "https://misterhr.vercel.app"],  # Add production URL later
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Pydantic models for request/response validation
class HealthResponse(BaseModel):
    """Health check response model."""
    status: str
    message: str
    version: str

class ResumeParseRequest(BaseModel):
    """Resume parsing request model."""
    file_url: str
    content_type: str  # "pdf" | "docx" | "text"

class ResumeParseResponse(BaseModel):
    """Resume parsing response model."""
    success: bool
    data: dict | None = None
    error: str | None = None

# Health check endpoint
@app.get("/", response_model=HealthResponse)
async def health_check():
    """Health check endpoint to verify API status."""
    return HealthResponse(
        status="healthy",
        message="MisterHR AI platform is running",
        version="1.0.0"
    )

# Resume parsing endpoint (placeholder for now)
@app.post("/api/parse-resume", response_model=ResumeParseResponse)
async def parse_resume(request: ResumeParseRequest):
    """Parse resume from multiple formats using ResumeParserAgent."""
    try:
        # TODO: Implement ResumeParserAgent integration
        # For now, return mock response with enhanced data structure
        return ResumeParseResponse(
            success=True,
            data={
                "personal_info": {
                    "name": "Mock User",
                    "email": "mock@example.com",
                    "phone": "+1-555-0123",
                    "title": "Senior Software Engineer",
                    "location": "San Francisco, CA"
                },
                "experience": [
                    {
                        "title": "Senior Software Engineer",
                        "company": "Tech Corp",
                        "duration": "2021-2023",
                        "years": 2,
                        "achievements": ["Led team of 5 developers", "Reduced deployment time by 60%"],
                        "technologies": ["Python", "React", "Docker"]
                    }
                ],
                "education": [
                    {
                        "degree": "Master of Computer Science",
                        "institution": "Stanford University",
                        "year": 2020,
                        "grade": "3.8 GPA"
                    }
                ],
                "skills": {
                    "technical": ["Python", "React", "Node.js", "PostgreSQL"],
                    "soft": ["Leadership", "Communication", "Problem Solving"],
                    "proficiency_levels": {
                        "Python": "expert",
                        "React": "advanced",
                        "PostgreSQL": "intermediate"
                    }
                },
                "projects": [
                    {
                        "name": "Cloud Migration Initiative",
                        "description": "Led migration from on-premises to cloud infrastructure",
                        "technologies": ["AWS", "Terraform", "Docker"],
                        "impact": "60% cost reduction, improved scalability"
                    }
                ],
                "certifications": [
                    {
                        "name": "AWS Certified Solutions Architect",
                        "issuer": "Amazon Web Services",
                        "year": 2022
                    }
                ],
                "online_presence": {
                    "github": "https://github.com/mockuser",
                    "linkedin": "https://linkedin.com/in/mockuser",
                    "portfolio": "https://mockportfolio.dev",
                    "verified": False
                }
            }
        )
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Resume parsing failed: {str(e)}"
        )

# Generate tailored resume endpoint (placeholder for now)
@app.post("/api/generate-resume")
async def generate_resume():
    """Generate tailored resume using ContentGeneratorAgent."""
    try:
        # TODO: Implement ContentGeneratorAgent integration
        return {"success": True, "message": "Resume generation initiated"}
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Resume generation failed: {str(e)}"
        )

# AI agent orchestration endpoint (placeholder for now)
@app.post("/api/batch-process")
async def batch_process():
    """Process multiple resumes using BatchProcessingAgent."""
    try:
        # TODO: Implement BatchProcessingAgent integration
        return {"success": True, "message": "Batch processing initiated"}
    except Exception as e:
        raise HTTPException(
            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
            detail=f"Batch processing failed: {str(e)}"
        )

@app.on_event("startup")
async def startup_event():
    """Initialize resources on startup."""
    print("🚀 Starting MisterHR Backend...")
    print("🔗 Agent orchestration system initializing...")
    # TODO: Initialize LangChain agents here
    print("✅ Backend ready for requests")

@app.on_event("shutdown")
async def shutdown_event():
    """Clean up resources on shutdown."""
    print("🛑 Shutting down MisterHR Backend...")
    # TODO: Clean up agent resources here
    print("✅ Backend shutdown complete")

if __name__ == "__main__":
    # For development mode
    port = int(os.getenv("PORT", 8000))
    host = os.getenv("HOST", "127.0.0.1")

    uvicorn.run(
        "main:app",
        host=host,
        port=port,
        reload=True,  # Enable auto-reload in development
        log_level="info"
    )
