requirements.txt
# MisterHR Backend Dependencies

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI & LangChain
langchain==0.1.0
langchain-openai==0.0.5
openai==1.3.5

# Data Processing
spacy==3.7.2
PyMuPDF==1.23.4
pandas==2.1.3
numpy==1.25.2

# Database & Supabase
supabase==2.0.3
psycopg2-binary==2.9.9

# File Processing
python-multipart==0.0.6
aiofiles==23.2.1

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Background Tasks & Redis (for future)
redis==5.0.1
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.37.1

# Environment & Configuration
python-dotenv==1.0.0
